pyproject.toml
[tool.poetry]
name        = "jurisCrew"
version     = "0.1.0"
description = "AI Juridical Assistant"
authors     = ["Fred Nesti <fredericonesti@gmail.com>"]

[build-system]
requires = ["setuptools>=46.1.0", "setuptools<=58.2.0", "setuptools_scm[toml]>=5", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python           = "3.10.15"
crewai           = "0.86.0" 
crewai-tools     = "0.17.0" 
python-dotenv    = "1.0.0"
streamlit        = "1.40.1"
chromadb         = "0.5.20"
pydantic         = "2.7.4"
langchain        = "0.3.6"
yarl             = "1.18.0"
spider-client    = "0.1.25"
selenium         = "4.27.1"
beautifulsoup4   = "4.12.3"

[logger]
level = "debug"

[browser]
serverAddress = "localhost"
serverPort = 8502

[tool.poetry.scripts]
start = "python -m streamlit run ./src/gui/app.py:main"


